azure-translate-text/requirements.txt
requests>=2.31.0
trafilatura>=1.8.0
python-dotenv>=1.0.1
